Test with M1 macOS CI runner
This adds macos-14 as a value for "os". The new macOS 14 runners
use Apple Silicon M1 systems (64-bit ARM CPUs), allowing GitPython
to be tested on ARM64 actions/python-versions builds of Python,
such as python-3.12.1-darwin-arm64.tar.gz.

https://github.blog/changelog/2024-01-30-github-actions-introducing-the-new-m1-macos-runner-available-to-open-source/
https://github.blog/changelog/2024-01-30-github-actions-macos-14-sonoma-is-now-available/

This commit doesn't exclude any `os`/`python-version` combinations,
even though not all versions of Python that GitPython supports are
currently available via the setup-python (and 3.7 builds are very
unlikely ever to be available for macos-14). This is to verify that
the currently unsupported versions are 3.7, 3.8, 3.9, and no
others. The next commit will exclude the unavailable versions.