Work around different ownership in container
This handles the "dubious ownership" error for the Alpine Linux
container using safe.directory as is done in the Cygwin job.

Another approach may be to actually use a limited user account in
the container, though, and that may be better, since I expect some
of the rmtree tests in test_util.py to fail due to the root user
being able to perform a delete operation the tests assume cannot be
done.