Suggest full-path refresh() in failure message
This does not suggest or recommend *preferring* to explicitly call
refresh() over the other other techniques, but it clarifies that
the use of refresh() being presented needs a path argument. It also
shows that path in the form of a full path, so users are less
likely to be misled into thinking a command name or other relative
path should be passed to refresh(), which should rarely be done
(since refresh(path) resolves path).