Undo a couple minor black-incompatible changes
Two of the docstrings in git.remote need to have a newline before
the conceptually "first" line in order for black to accept the
indentation of the subsequent text, which appears intentional and
works fine as reStructuredText. Otherwise black dedents the code
block, which affects rendering and is also less readable.