Better clarify Submodule.branch_path documentation
This revisits the changes in 3813bfb and takes a different
approach, using the phrase "full repository-relative path" that is
used elsewhere in the documentation and seems clear in context.

Although this brings back the potential confusion around having the
terms "full" and "relative" used together to describe a path, this
may no longer be an issue now that the phrase "repository-relative"
is used here as it is elsewhere.

(In particular, see the SymbolicReference.to_full_path docstring.)