Fix exception type in require_remote_ref_path docstring
The require_remote_ref_path decorator has always used ValueError
(ever since its introduction in a92ab80) but was documented as
using TypeError.

ValueError is the correct exception to raise here, since this is
not any kind of type error or related condition. So the bug wasn't
in the function's behavior, but instead in the way that behavior
was documented. (The bug was fairly minor, since the function is
not listed in __all__.)