Fix possible inaccuracy in Lit_config_levels docstring
These are not necessarily what one means by the "scope" of a
configuration variable, in part because of the nature of the subtle
distinction between "user" and "global", and in part because of
other issues such as how setting a variable for a single command
with "-c" has a scope which is not listed.

This also brings the docstring somewhat more in line with how these
values are documented elsewhere in GitPython.