Use phrases like "git object type" where applicable
In some recently introduced or expanded docstrings, I had overused
phrases like "kind of git object" with the hope of avoiding
confusion with the meanings of "type" relevant to Python (i.e.,
"class" or "static type"). But this made the relationship to git's
own notion of "object type" less clear than it could be, especially
in docstrings that also included links to the gitglossary(7) entry
for "object type" (because those links' relevance was less clear).

This dials back the use of "kind" to where I am more confident that
it is clarifying or at least not confusing.