Don't bind unused _assertion_msg_format
The _assertion_msg_format module attribute (global variable) of
git.objects.base was formerly used in an assertion check that has
since been commented out but not completely removed.

It may be that both it and the commented-out code that uses it
should simply be removed (they will be in the git history, after
all), but this change just brings them in line by also commenting
out the variable.