Add return-type annotation on __init__ methods
Although sometimes found unintuitive, the reutrn type of a class's
__init__ method is best annotated as None, as in other functions
that return implicitly or by operand-less return statements.

Note that this is only a minor improvement, effectively just a
style fix, because mypy treats __init__ specially and, *when at
least one parameter is annotated*, its return type is implicitly
None rather than implicitly Any like other functions. All the
__init__ methods modified here did already have one or more
annotated parameters.

However, having __init__ methods without the return type makes it
easier to introduce a bug where an __init__ method with no
parameters besides self -- which itself should almost always be
unannotated and is properly inferred -- is written and the return
annotation needed to get mypy to regard it as statically typed,
so it checks it at all, is omitted. (It is also inelegant when
one considers the meaning of __init__ and the distinction between
it and __new__.)

This commit does not add any return type annotations to functions
in the test suite, since the test suite doesn't currently use
static typing.

Further reading:

- https://peps.python.org/pep-0484/#the-meaning-of-annotations
- https://github.com/python/mypy/issues/604
- https://github.com/gitpython-developers/GitPython/pull/1755#issuecomment-1837419098