Fix new mypy confusion about kill_after_timeout type
The refactoring in 1ef3365 before this added a new mypy error on
non-Windows platforms, where mypy failed to infer that the type of
kill_after_timeout was `float` (rather than `float | None`) at the
point in the code where it was used as a captured variable in the
newly introduced communicate() helper.

This was even though the variable is never rebound there or in the
enclosing scope that introduced it. So introducing and using a new
variable that holds the same reference, which is sufficient to fix
the problem and is the approach taken here, is not a behavioral
change.